% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_percentile_bands}
\alias{plot_percentile_bands}
\title{Plot percentile bands for wealth trajectories}
\usage{
plot_percentile_bands(wealth)
}
\arguments{
\item{wealth}{matrix from simulate_wealth}
}
\value{
ggplot object
}
\description{
Plot percentile bands for wealth trajectories
}
