% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{sp500_annual_data}
\alias{sp500_annual_data}
\title{Get S&P 500 Annual Data}
\usage{
sp500_annual_data()
}
\value{
A data frame with columns 'Year' and 'Return'.
}
\description{
Returns the full historical S&P 500 annual returns data frame.
}
