% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp500_data.R
\name{sp500_annual_sample}
\alias{sp500_annual_sample}
\title{Load S&P 500 annual returns sample data}
\usage{
sp500_annual_sample()
}
\value{
A numeric vector of annual returns
}
\description{
Load S&P 500 annual returns sample data
}
