% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_and_summaries.R
\name{summarize_simulations}
\alias{summarize_simulations}
\title{Summarize multiple simulations as a tibble}
\usage{
summarize_simulations(wealth)
}
\arguments{
\item{wealth}{matrix returned by simulate_wealth}
}
\description{
Summarize multiple simulations as a tibble
}
