% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp500_data.R
\name{sp500_annual_data}
\alias{sp500_annual_data}
\title{Get full S&P 500 data with years}
\usage{
sp500_annual_data()
}
\value{
A data frame with year and return columns
}
\description{
Get full S&P 500 data with years
}
