% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R, R/sp500_annual_sample.R
\name{sp500_annual_sample}
\alias{sp500_annual_sample}
\title{Get S&P 500 Annual Returns Sample Vector}
\usage{
sp500_annual_sample()

sp500_annual_sample()
}
\value{
A numeric vector of historical returns.

A numeric vector of historical returns.
}
\description{
Returns a numeric vector of historical S&P 500 annual returns.

Returns a numeric vector of historical S&P 500 annual returns.
}
